Attempt to fix #2
This isn't actually tested in place because I can't compile csm from source.
However, I tested the code snippet separately and it reads all the
molecules from a PDB file correctly.

One thing that should be checked is the behavior of csm when reading
a PDB file with 0 or 1 molecules.